# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bodapay API
  description: |
    REST API for the Bodapay cross-border payment platform built on Stellar.

    ## Authentication
    All protected endpoints require a Bearer JWT token in the `Authorization` header.
    Obtain a token via `POST /api/v1/auth/login`.

    ## Versioning
    The API supports two versions:
    - **v1** — deprecated, still functional
    - **v2** — current, recommended

    Set the desired version via the `Accept-Version` header or use the versioned path prefix (`/api/v1/` or `/api/v2/`).
  version: "2.0.0"
  contact:
    name: Bodapay Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development (v1)
  - url: http://localhost:8080/api/v2
    description: Local development (v2)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request body"
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          minLength: 8
          example: "s3cur3P@ss!"
        first_name:
          type: string
          example: Alice
        last_name:
          type: string
          example: Smith
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          example: "s3cur3P@ss!"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIs..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIs..."
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 42
        email:
          type: string
          example: alice@example.com
        first_name:
          type: string
          example: Alice
        last_name:
          type: string
          example: Smith
        role:
          type: string
          example: user
        created_at:
          type: string
          format: date-time
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: alice@example.com
              password: "s3cur3P@ss!"
              first_name: Alice
              last_name: Smith
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Obtain JWT access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...